Refactor: misc
